MCP_demo/pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph==0.4.8",
    "mcp>=1.12.2",
    "python-dotenv>=1.1.1",
]
